Fix the testsuite
As PHPUnit 10+ does not include a builtin Prophecy integration, we
should not rely on it to write tests.

Thus, when writing tests for Prophecy itself, it is better if the test
shows usages of the Prophecy API instead of hiding it in the PHPunit
integration (which would also run additional code to verify predictions,
which might make the test confusing).
